move writers to its own module
