+++ @xen-orchestra/backups/writers/_AbstractDeltaWriter.js
