+++ @xen-orchestra/backups/writers/_AbstractFullWriter.js
