+++ @xen-orchestra/backups/writers/_DeltaBackupWriter.js
@@ -7,13 +7,14 @@ const { chainVhd, checkVhdChain, default: Vhd } = require('vhd-lib')
 const { createLogger } = require('@xen-orchestra/log')
 const { dirname } = require('path')
 
+const { formatFilenameDate } = require('../_filenameDate')
+const { getOldEntries } = require('../_getOldEntries')
+const { getVmBackupDir } = require('../_getVmBackupDir')
+const { Task } = require('../Task')
+
 const { AbstractDeltaWriter } = require('./_AbstractDeltaWriter')
 const { checkVhd } = require('./_checkVhd')
-const { formatFilenameDate } = require('./_filenameDate')
-const { getOldEntries } = require('./_getOldEntries')
-const { getVmBackupDir } = require('./_getVmBackupDir')
 const { packUuid } = require('./_packUuid')
-const { Task } = require('./Task')
 
 const { warn } = createLogger('xo:backups:DeltaBackupWriter')
