+++ @xen-orchestra/backups/writers/_FullBackupWriter.js
@@ -1,9 +1,10 @@
+const { formatFilenameDate } = require('../_filenameDate')
+const { getOldEntries } = require('../_getOldEntries')
+const { getVmBackupDir } = require('../_getVmBackupDir')
+const { isValidXva } = require('../isValidXva')
+const { Task } = require('../Task')
+
 const { AbstractFullWriter } = require('./_AbstractFullWriter')
-const { formatFilenameDate } = require('./_filenameDate')
-const { getOldEntries } = require('./_getOldEntries')
-const { getVmBackupDir } = require('./_getVmBackupDir')
-const { isValidXva } = require('./isValidXva')
-const { Task } = require('./Task')
 
 exports.FullBackupWriter = class FullBackupWriter extends AbstractFullWriter {
   constructor(backup, remoteId, settings) {
