+++ @xen-orchestra/backups/writers/_checkVhd.js
