+++ @xen-orchestra/backups/writers/_packUuid.js
