+++ @xen-orchestra/backups/writers/index.js
@@ -1,0 +1,4 @@
+exports.ContinuousReplicationWriter = require('./_ContinuousReplicationWriter').ContinuousReplicationWriter
+exports.DeltaBackupWriter = require('./_DeltaBackupWriter').DeltaBackupWriter
+exports.DisasterRecoveryWriter = require('./_DisasterRecoveryWriter').DisasterRecoveryWriter
+exports.FullBackupWriter = require('./_FullBackupWriter').FullBackupWriter
